Share MCP plan-analysis JSON formatter across Lite/Dashboard
The byte-identical BuildAnalysisResult + CollectNodes (parse plan XML -> analyze
-> serialize structured JSON) lived privately in both apps' McpPlanTools. Moved
to PerformanceMonitor.PlanAnalysis.McpPlanAnalysisFormatter; both apps' five MCP
tool wrappers (which fetch the plan XML from SQL Server vs DuckDB) now call the
shared formatter.

Required adding a PlanAnalysis -> Common project reference (the formatter uses
McpHelpers.Truncate/JsonOptions) and granting Common's InternalsVisibleTo to
PerformanceMonitor.PlanAnalysis. No dependency cycle: Common references no other
project. Dropped now-unused usings from both McpPlanTools files.

Pure refactor, no behavior change. Solution builds clean; Dashboard.Tests (488),
Lite.Tests (544), Installer.Tests (80) all pass.
